Implement initialization of sweep queues
This adds several new facilities, most notably the ability to iterate over all leaf boxes. This also adds information about the splitting edges (e.g., from the golden section split). This was necessary because sweeps allow the level of a box to be incremented without changing its boundaries, and hence the split edge can't be computed from the level.